v9.8

+++ customize.sh
@@ -161,37 +161,68 @@ fi
 rm -rf $MODPATH/system_10
 
 # function
-run_check_function() {
-LISTS=`strings $MODPATH/system/vendor$DIR/$DES | grep ^lib | grep .so`
-FILE=`for LIST in $LISTS; do echo $SYSTEM$DIR/$LIST; done`
-ui_print "- Checking"
-ui_print "$NAME"
-ui_print "  function at"
-ui_print "$FILE"
-ui_print "  Please wait..."
-if ! grep -q $NAME $FILE; then
-  ui_print "  Function not found."
-  ui_print "  Replaces /system$DIR/$LIB"
-  mv -f $MODPATH/system_support$DIR/$LIB $MODPATH/system$DIR
-fi
-ui_print " "
-}
 check_function() {
-if [ "$IS64BIT" == true ]; then
-  DIR=/lib64
-  run_check_function
+if [ -f $MODPATH/system_support$DIR/$LIB ]; then
+  ui_print "- Checking"
+  ui_print "$NAME"
+  ui_print "  function at"
+  ui_print "$FILE"
+  ui_print "  Please wait..."
+  if ! grep -q $NAME $FILE; then
+    ui_print "  Function not found."
+    ui_print "  Replaces /system$DIR/$LIB."
+    mv -f $MODPATH/system_support$DIR/$LIB $MODPATH/system$DIR
+    [ "$MES" ] && ui_print "$MES"
+  fi
+  ui_print " "
 fi
-if [ "$LIST32BIT" ]; then
-  DIR=/lib
-  run_check_function
-fi
+}
+find_file() {
+for LIB in $LIBS; do
+  if [ -f $MODPATH/system_support$DIR/$LIB ]; then
+    FILE=`find $SYSTEM$DIR $SYSTEM_EXT$DIR -type f -name $LIB`
+    if [ ! "$FILE" ]; then
+      ui_print "- Using /system$DIR/$LIB."
+      mv -f $MODPATH/system_support$DIR/$LIB $MODPATH/system$DIR
+      ui_print " "
+    fi
+  fi
+done
 }
 
 # check
 NAME=_ZN7android8hardware23getOrCreateCachedBinderEPNS_4hidl4base4V1_05IBaseE
 DES=vendor.dolby.hardware.dms@2.0.so
 LIB=libhidlbase.so
-check_function
+MES="  Dolby Atmos may not work."
+if [ "$IS64BIT" == true ]; then
+  DIR=/lib64
+  LISTS=`strings $MODPATH/system/vendor$DIR/$DES | grep ^lib | grep .so`
+  FILE=`for LIST in $LISTS; do echo $SYSTEM$DIR/$LIST; done`
+  check_function
+fi
+if [ "$LIST32BIT" ]; then
+  DIR=/lib
+  LISTS=`strings $MODPATH/system/vendor$DIR/$DES | grep ^lib | grep .so`
+  FILE=`for LIST in $LISTS; do echo $SYSTEM$DIR/$LIST; done`
+  check_function
+fi
+
+# check
+if [ "$SYSTEM_10" == true ]; then
+  LIBS="libhidltransport.so libhwbinder.so"
+else
+  LIBS=libhidltransport.so
+fi
+if [ "$IS64BIT" == true ]; then
+  DIR=/lib64
+  find_file
+fi
+if [ "$LIST32BIT" ]; then
+  DIR=/lib
+  find_file
+fi
+rm -rf $MODPATH/system_support
 
 # sepolicy
 FILE=$MODPATH/sepolicy.rule
@@ -279,10 +310,10 @@ done
 
 # conflict
 if [ "`grep_prop dolby.mod $OPTIONALS`" == 1 ]; then
-  NAMES="dolbyatmos DolbyAudio MotoDolby
+  NAMES="dolbyatmos DolbyAudio MotoDolby DolbyAtmosSP
          DolbyAtmos360"
 else
-  NAMES="dolbyatmos DolbyAudio MotoDolby
+  NAMES="dolbyatmos DolbyAudio MotoDolby DolbyAtmosSP
          DolbyAtmos360 dsplus Dolby"
 fi
 conflict
@@ -502,24 +533,14 @@ else
   EIM=false
 fi
 }
-run_find_file() {
-for NAME in $NAMES; do
-  FILE=`find $SYSTEM$DIR $SYSTEM_EXT$DIR -type f -name $NAME`
-  if [ ! "$FILE" ]; then
-    ui_print "- Using /system$DIR/$NAME"
-    cp -f $MODPATH/system_support$DIR/$NAME $MODPATH/system$DIR
-    ui_print " "
-  fi
-done
-}
-find_file() {
-if [ "$IS64BIT" == true ]; then
-  DIR=/lib64
-  run_find_file
-fi
-if [ "$LIST32BIT" ]; then
-  DIR=/lib
-  run_find_file
+eim_cache_warning() {
+if echo $EIMDIR | grep -q cache; then
+  ui_print "  Please do not ever wipe your /cache"
+  ui_print "  as long as this module is installed!"
+  ui_print "  If your /cache is wiped for some reasons,"
+  ui_print "  then you need to uninstall this module and reboot first,"
+  ui_print "  then reinstall this module afterwards"
+  ui_print "  to get this module working correctly."
 fi
 }
 patch_manifest_eim() {
@@ -534,14 +555,14 @@ if [ $EIM == true ]; then
     fi
     if ! grep -A2 vendor.dolby.hardware.dms $DES | grep -q 2.0; then
       ui_print "- Patching"
-      ui_print "$SRC"
-      ui_print "  systemlessly using early init mount..."
+      ui_print "$DES"
       sed -i '/<manifest/a\
     <hal format="hidl">\
         <name>vendor.dolby.hardware.dms</name>\
         <transport>hwbinder</transport>\
         <fqname>@2.0::IDms/default</fqname>\
     </hal>' $DES
+      eim_cache_warning
       ui_print " "
     fi
   else
@@ -561,10 +582,10 @@ if [ $EIM == true ]; then
     fi
     if ! grep -Eq 'u:object_r:hal_dms_hwservice:s0|u:object_r:default_android_hwservice:s0' $DES; then
       ui_print "- Patching"
-      ui_print "$SRC"
-      ui_print "  systemlessly using early init mount..."
+      ui_print "$DES"
       sed -i '1i\
 vendor.dolby.hardware.dms::IDms u:object_r:hal_dms_hwservice:s0' $DES
+      eim_cache_warning
       ui_print " "
     fi
   else
@@ -591,16 +612,6 @@ remount_rw
 
 # early init mount dir
 early_init_mount_dir
-
-# check
-chcon -R u:object_r:system_lib_file:s0 $MODPATH/system_support/lib*
-if [ "$SYSTEM_10" == true ]; then
-  NAMES="libhidltransport.so libhwbinder.so"
-else
-  NAMES=libhidltransport.so
-fi
-find_file
-rm -rf $MODPATH/system_support
 
 # patch manifest.xml
 FILE="$INTERNALDIR/mirror/*/etc/vintf/manifest.xml
@@ -725,9 +736,9 @@ if [ "$BOOTMODE" == true ]\
   pm grant $PKG $NAME
   if ! dumpsys package $PKG | grep -q "$NAME: granted=true"; then
     ui_print "  ! Failed."
-    ui_print "  Maybe insufficient storage."
-    ui_print "  or maybe there is in-built $PKG exist."
-    ui_print "  Just ignore this."
+    ui_print "    Maybe insufficient storage"
+    ui_print "    or maybe there is in-built $PKG exist."
+    ui_print "    Just ignore this."
   fi
   RES=`pm uninstall -k $PKG 2>/dev/null`
   ui_print " "
